[CCTRI-1588] Connect Save module type json functionality to Module Maker UI

+++ src/container/Content/components/Properties/components/SupportedAPI/SupportedAPI.js
@@ -14,15 +14,20 @@ class SupportedAPI extends React.Component {
     }
 
     onCheckboxToggle = (event) => {
+
         this.props.updateSupportedAPI(this.state.title);
 
         if (event.target.checked) {
-        this.props.addCapability(this.state.class);
+            this.props.addCapability(this.state.class);
         }
+
         else {
-        this.props.deleteCapability(this.state.class);
+            if (this.props.syncProperties["supported-apis"].filter(
+                elem => elem.startsWith(this.state.class)).length === 1) {
+                this.props.deleteCapability(this.state.class);
+            }
+          }
         }
-    }
 
     render() {
         return (
@@ -40,4 +45,10 @@ const mapDispatchToProps = {
     updateSupportedAPI
 }
 
-export default connect(null, mapDispatchToProps)(SupportedAPI);
+const mapStateToProps = state => {
+    return {
+        syncProperties: state.properties
+    }
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(SupportedAPI);
